% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-helpers.R
\name{ConstructVarAnnotate}
\alias{ConstructVarAnnotate}
\title{ConstructVarAnnotate}
\usage{
ConstructVarAnnotate(FCS.file.name)
}
\arguments{
\item{FCS.file.name}{The full file path to one FCS file intended for analysis}
}
\value{
List mapping channel names (taken from the "name" field of a
flowFrame) to alternative names from the flowFrame "desc" fields
}
\description{
\code{ConstructVarAnnotate} returns a list that can be used during FLOWMAPR
analysis to change channel names from the default "name" field used in a
flowFrame to the names provided in the "desc" fields. Generally in CyTOF/mass
cytometry or flow cytometry data, the original "name" field corresponds to the
measurement channel (e.g. FITC or Sm154Di) while the "desc" field will correspond
to the name of the marker entered by the user at the machine (e.g. cCaspase3 or
CyclinB1).
}
\examples{
ConstructVarAnnotate(FCS.file.name = "Desktop/A.fcs")
}
