% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-graphs.R
\name{ConvertToGraphML}
\alias{ConvertToGraphML}
\title{Saving an igraph object as a local graphml file.}
\usage{
ConvertToGraphML(output.graph, file.name)
}
\arguments{
\item{output.graph}{An igraph graph object to be saved as a local file.}

\item{file.name}{A short, descriptive phrase for the file name of
the saved igraph graph object. The user need not include the ".graphml"
file extension as that is automatically appended.}
}
\description{
\code{ConvertToGraphML} saves a copy of an igraph graph object
as a local graphml file in the current file directory with the
file name containing the current date, a supplied string in
\code{file.name}, and the current time.
}
\examples{
current.graph <- igraph::sample_k_regular(no.of.nodes = 10, k = 2, directed = FALSE, 
multiple = FALSE)
file.name <- "Practice_Graph"
\dontrun{ConvertToGraphML(current.graph, file.name)}
}
