% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{DownsampleFCS}
\alias{DownsampleFCS}
\title{SPADE: density-dependent downsampling}
\usage{
DownsampleFCS(fcs.file.names, clustering.var, channel.annotate,
  channel.remove, exclude.pctile = 0.01, target.pctile = 0.99,
  target.number = NULL, target.percent = 0.1, transform = TRUE)
}
\arguments{
\item{fcs.file.names}{A vector of full file paths to the FCS files to be used}

\item{clustering.var}{Vector naming channels to be used to calculate distances/differences
between cells for clustering (if requested) and edge-drawing steps}

\item{channel.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{channel.remove}{Vector naming channels to be removed from all loaded FCS data}

\item{exclude.pctile}{Optional variable, numeric value for the downsampling_exclude_pctile variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.pctile}{Optional variable, numeric value for the downsampling_target_pctile variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.number}{Optional variable, numeric value for the downsampling_target_number variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.percent}{Optional variable, numeric value for the downsampling_target_percent variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{transform}{Logical specifying whether to transform the data using an Asinh
transform typical of CyTOF/mass cytometry datasets, default is set to \code{TRUE}}
}
\description{
This code was adapted from the spade R package,
available here: \url{https://github.com/nolanlab/spade}.
}
\details{
Specifically, the code in this file is adapted from downsample.R.

This code, authored by M. Linderman, P. Qiu, E. Simonds, Z. Bjornson,
and maintained by Michael Linderman <michael.d.linderman@gmail.com>,
was used under the GNU General Public License v. 2.0
(\url{https://github.com/nolanlab/spade/blob/master/LICENSE}) available
here: \url{https://opensource.org/licenses/GPL-2.0}. In accordance with these
license rules, our code is available under GPL-3.0.

UPDATE 12/17/19: Spade is no longer maintained
We have included relevant function here to maintain functionality of FLOWMAP
}
