% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{flowMap-package}
\alias{flowMap-package}
\title{Flow cytometry data cross-sample comparison}
\description{
This package implements a method for matching cell populations across multiple flow cytometry samples.
}
\author{
Chiaowen Hsiao \email{joyce.hsiao1@gmail.com}
}

