% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FLOWMAP}
\alias{FLOWMAP}
\title{FLOWMAP - generate FLOWMAPR analysis results from FCS files in R}
\usage{
FLOWMAP(mode = c("single", "multi", "one", "static-multi"),
  var.remove = c(), var.annotate = NULL, clustering.var,
  cluster.numbers = 100, cluster.mode = "hclust",
<<<<<<< HEAD
  distance.metric = "manhattan", umap.settings, files,
  density.metric = c("kNN", "radius"), minimum = 2, maximum = 5,
  save.folder = getwd(), subsamples = 200, name.sort = TRUE,
  downsample = FALSE, seed.X = 1, savePDFs = TRUE,
  graph.out = c("ForceDirected"), which.palette = "bluered",
=======
  distance.metric = "manhattan", files, density.metric = c("kNN",
  "radius"), minimum = 2, maximum = 5, save.folder = getwd(),
  subsamples = 200, name.sort = TRUE, downsample = FALSE,
  seed.X = 1, savePDFs = TRUE, which.palette = "bluered",
>>>>>>> 7017d079511d4db1eae54b327920cfb0e134034f
  exclude.pctile = NULL, target.pctile = NULL, target.number = NULL,
  target.percent = NULL, k = 10, per = 5, ...)
}
\arguments{
\item{mode}{FLOWMAPR mode to use in analysis based on starting input,
available options include \code{c("single", "multi", "one", "static-multi")}}

\item{var.remove}{Vector naming channels to be removed from all downstream analysis, default
value is an empty vector, meaning that no channels will be removed}

\item{var.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data, default value is \code{NULL}, which will then autogenerate the
\code{var.annotate} values using the \code{ConstructVarAnnotate()} function}

\item{clustering.var}{Vector naming channels to be used to calculate distances/differences
between cells for clustering (if requested) and edge-drawing steps}

\item{cluster.numbers}{A single numeric or a vector of numerics specifying how many clusters
to generate from each separate FCS file}

\item{cluster.mode}{Character specifying which clustering algorithm to use, valid options include \code{c("hclust", "kmeans")}}

\item{distance.metric}{Character specifying which metric to use to calculate between-node distances, valid options include \code{c("manhattan", "euclidean")}}

<<<<<<< HEAD
\item{umap.settings}{TBD}

=======
>>>>>>> 7017d079511d4db1eae54b327920cfb0e134034f
\item{files}{File paths for FCS files to be used or a folder containing
the FCS files to be used in analysis}

\item{density.metric}{Character string specifying which method to use for local density estimation for edge assignment in graph building}

\item{minimum}{Numeric value specifying the minimum number of edges that will be allotted
during each density-dependent edge-building step of the FLOW-MAP graph, default value is
set to \code{2}, no less than 2 is recommended}

\item{maximum}{Numeric value specifying the maximum number of edges that will be allotted
during each density-dependent edge-building step of the FLOW-MAP graph, default value is
set to \code{5}, no less than 3 is recommended}

\item{save.folder}{Directory where all results generated should be saved}

\item{subsamples}{A single numeric or a vector of numerics specifying how many cells to
subsample from each FCS file}

\item{name.sort}{Logical specifying whether to sort FCS file path names alphanumerically or use
them in the order supplied by the user}

\item{downsample}{Logical specifying whether to use SPADE density-dependent downsampling}

\item{seed.X}{Numeric value for the seed to set for reproducible FLOWMAPR runs}

\item{savePDFs}{Logical specifying whether to generate PDFs for the resolved graph with
nodes colored by each parameter}

\item{graph.out}{TBD}

\item{which.palette}{Optional variable, character specifying which color palette to use
in generated PDFs, valid options include \code{c("bluered", "jet", "CB")}, where \code{"CB"}
is a colorblind-friendly option}

\item{exclude.pctile}{Optional variable, numeric value for the downsampling_exclude_pctile variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.pctile}{Optional variable, numeric value for the downsampling_target_pctile variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.number}{Optional variable, numeric value for the downsampling_target_number variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{target.percent}{Optional variable, numeric value for the downsampling_target_percent variable
used as described in the SPADE driver function, see the documentation for the spade package at
\url{https://github.com/nolanlab/spade}}

\item{umap.k}{TBD}
}
\value{
the force-directed layout resolved igraph graph object
}
\description{
\code{FLOWMAP} generates FLOWMAPR analysis results from FCS files. For more information,
as well as a guide for how to choose the best settings for your analysis, go to
our GitHub repo \url{https://github.com/zunderlab/FLOWMAP/}.
}
