% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-alt.R
\name{FLOWMAPfromDF}
\alias{FLOWMAPfromDF}
\title{FLOWMAPfromDF - generate FLOWMAPR analysis results from
dataframes in R}
\usage{
FLOWMAPfromDF(mode = c("single", "multi", "one"), df, project.name,
  clustering.var, distance.metric = "manhattan", minimum = 2, maximum = 5,
  per = 1, save.folder = getwd(), time.col.label = "Time",
  condition.col.label = NULL, name.sort = TRUE, clustering = FALSE,
  seed.X = 1, savePDFs = TRUE, which.palette = "bluered",
  cluster.numbers = NULL)
}
\arguments{
\item{mode}{???}

\item{df}{???}

\item{project.name}{???}

\item{clustering.var}{???}

\item{distance.metric}{???}

\item{minimum}{???}

\item{maximum}{???}

\item{per}{???}

\item{save.folder}{???}

\item{time.col.label}{???}

\item{condition.col.label}{???}

\item{name.sort}{???}

\item{clustering}{???}

\item{seed.X}{???}

\item{savePDFs}{???}

\item{which.palette}{???}

\item{cluster.numbers}{???}
}
\value{
???

\url{http://en.wikipedia.org/}
}
\description{
\code{FLOWMAPfromDF} returns the ???
}
\details{
This function ???
}
\examples{
FLOWMAPfromDF()

\dontrun{
FLOWMAPfromDF()
}
}
