\docType{package}
\name{flowMatch-package}
\alias{flowMatch-package}
\title{Flow cytometry data cross-sample comparison}
\description{
  This package implements a method for matching cell
  populations across multiple flow cytometry samples.
}
\author{
  Chiaowen Joyce Hsiao \email{joyce.hsiao1@gmail.com}
}

