% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FRstats-class}
\alias{FRstats-class}
\title{FR statistics and p-values generated from one single draw of the population pair comparison}
\description{
This class stores the statistics required to compute median FR statistics across random draws.
}
\section{FIXME}{

## usage
## Accessors
getFRstats(object)
getPnorm(object)
}
\examples{
## see vignettes
}
\author{
Chiaowen Joyce Hsiao \email{joyce.hsiao1@gmail.com}
}

