% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{GetFCSNames}
\alias{GetFCSNames}
\title{GetFCSNames}
\usage{
GetFCSNames(folder, sort = TRUE)
}
\arguments{
\item{folder}{the full path for a folder that contains subfolders, each
containing the FCS files from a specific timepoint to be used}

\item{sort}{Logical specifying whether FCS files should be sorted in alphanumeric order}
}
\value{
Vector with the file paths of FCS files found in the given folder
}
\description{
\code{GetFCSNames} returns the FCS file names from a given folder.
}
\examples{
folder <- "Desktop/FCS_Files"
GetFCSNames(folder, sort = TRUE)
}
