% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFR}
\alias{getFR}
\title{FR test for a single cell population pair comparison}
\usage{
getFR(xx1, xx2)
}
\arguments{
\item{xx1}{events of a single cell population, organized in a matrix or a data.frame of events (rows) by
features (columns).}

\item{xx2}{events of a single cell population, organized in a matrix or a data.frame of events (rows) by
         features (columns).}
}
\value{
ww FR statistic.

runs number of within-group subtrees (large number of runs indicate high degree of dissimilarity
             between the two cell populations being compared.

mu expected number of runs (when the two cell populations are similarly distributed).

sigma2 variance of runs.

pNorm p-values of the FR statitic assuming large sample asymptotic normal assumption.
}
\description{
Compute FR statistic for any two cell populations in flow cytometry data. Runtime of the FR test is
quadratic in the number of events (nodes) totaling a single population pair comprison.
}
\examples{
## see vignettes
}
\author{
Chiaowen Joyce Hsiao \email{joyce.hsiao1@gmail.com}
}

