% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{GetMultiFCSNames}
\alias{GetMultiFCSNames}
\title{GetMultiFCSNames}
\usage{
GetMultiFCSNames(folder, sort = TRUE)
}
\arguments{
\item{folder}{the full path for a folder that contains the FCS files to be used}

\item{sort}{Logical specifying whether FCS files should be sorted in alphanumeric order}
}
\value{
List with each member containing a vector with the file paths
of FCS files found in the given subfolder
}
\description{
\code{GetFCSNames} returns the FCS file names as a list from a given folder,
that contains subfolders corresponding to each timepoint.
}
\examples{
folder <- "Desktop/FCS_Files"
GetMultiFCSNames(folder, sort = TRUE)
}
