% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-GUI.R
\name{launch_GUI}
\alias{launch_GUI}
\title{GUI for launching shiny APP}
\usage{
launch_GUI()
}
\arguments{
\item{message}{A message to determine if open the shiny APP}

\item{dir}{Result direcroty.}
}
\description{
A shiny APP for interactive exploration of the analysis results
}
\examples{
# launch_GUI()
}
