% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-GUI.R
\name{LaunchGUI}
\alias{LaunchGUI}
\title{GUI for launching shiny Application}
\usage{
LaunchGUI()
}
\arguments{
\item{message}{A message to determine if open the shiny Application}

\item{dir}{Result direcroty.}
}
\description{
This code was adapted from the cytofkit R package,
available here: https://github.com/JinmiaoChenLab/cytofkit/.
}
\details{
Specifically, the code in this file is adapted from cytofkit_GUI.R, but
modified to take user-specified inputs for FLOW-MAP runs and
then interface with the FLOWMAPR library back-end.

This code, authored and maintained by Jinmiao Chen and Hao Chen,
was used under the Artistic License 2.0 or Artistic-2.0 available
here: https://opensource.org/licenses/Artistic-2.0.
}
\examples{
# LaunchGUI()
}
