% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{LoadCleanFCS}
\alias{LoadCleanFCS}
\title{LoadCleanFCS}
\usage{
LoadCleanFCS(
  fcs.file.names,
  channel.remove,
  channel.annotate,
  subsamples = 1000,
  transform = TRUE
)
}
\arguments{
\item{fcs.file.names}{A vector of full file paths to the FCS files to be used}

\item{channel.remove}{Vector naming channels to be removed from all loaded FCS data}

\item{channel.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{subsamples}{Numeric or vector of numerics specifying how many cells to sample
from each FCS file, default is set to \code{1000}}

\item{transform}{Logical specifying whether to transform the data using an Asinh
transform typical of CyTOF/mass cytometry datasets, default is set to \code{TRUE}}
}
\value{
a list where each member is a dataframe containing the single-cell data
}
\description{
\code{LoadCleanFCS} returns a list of FCS file data, where each member is
a dataframe from a different timepoint. In the mode \code{"single"}, the list has
only a single element, a dataframe from the one FCS file path provided.
}
\examples{
fcs.file.names <- c("Desktop/A.fcs", "Desktop/B.fcs")
var.remove <- c("Channel3", "Channel4")
var.annotate <- list("c1" = "Channel1", "c2" = "Channel2",
"c3" = "Channel3", "c4" = "Channel4")

LoadCleanFCS(fcs.file.names, var.remove, var.annotate, subsamples = 100, transform = TRUE)
}
