% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{LoadMultiCleanFCS}
\alias{LoadMultiCleanFCS}
\title{LoadMultiCleanFCS}
\usage{
LoadMultiCleanFCS(list.of.file.names, channel.remove, channel.annotate,
  subsamples = 1000, transform = TRUE)
}
\arguments{
\item{list.of.file.names}{A list where each member contains a vector of full file paths
to the FCS files to be used}

\item{channel.remove}{Vector naming channels to be removed from all loaded FCS data}

\item{channel.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{subsamples}{Numeric or vector of numerics specifying how many cells to sample
from each FCS file, default is set to \code{1000}}

\item{transform}{Logical specifying whether to transform the data using an Asinh
transform typical of CyTOF/mass cytometry datasets, default is set to \code{TRUE}}
}
\value{
a list where each member is a list containing multiple dataframes containing
single-cell data from the same timepoint, but multiple conditions
}
\description{
\code{LoadMultiCleanFCS} returns a list of FCS file data, where each member is
a list that contains dataframe from different conditions from different timepoints.
}
\examples{
list.of.file.names <- list("1" = c("Desktop/A-1.fcs", "Desktop/B-1.fcs"),
"2" = c("Desktop/A-2.fcs", "Desktop/B-2.fcs"))
var.remove <- c("Channel3", "Channel4")
var.annotate <- list("c1" = "Channel1", "c2" = "Channel2",
"c3" = "Channel3", "c4" = "Channel4")

\dontrun{LoadMultiCleanFCS(list.of.file.names, var.remove, var.annotate, subsamples = 100, 
transform = TRUE)}
}
