% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDistmat}
\alias{makeDistmat}
\title{Generate a similarity matrix quantifying similarity between cell populations across flow cytmetry samples}
\usage{
makeDistmat(samples, sampleMethod = "proportional", sampleSize = 200,
  ndraws = 200)
}
\arguments{
\item{samples}{a list of data.frames or matrices of flow cytometry samples. Each data.frame or matrix
consists of columns of features (flow cytometry channels), followed by a column of population
      membership (\emph{id}).}

\item{sampleMethod}{downsampling method, options include \emph{equalSize} or \emph{proportional}. Both methods
sample events without replacement from the combined events in a single cell population pair
comparison. Using \emph{equalSize}, each sample includes an equal number of events from the
two cell populations being compared. Using \emph{proportional}, the ratio of the event
membership is same as the ratio of event membershiop prior to sampling.
(default: \emph{proportional})}

\item{sampleSize}{specifies \emph{S},the number of events to be included in each sample. For \emph{equalSize} sampling
, S/2 is sampled from each population. For \emph{proportional} sampling, the ratio of event membership
is the same as the ratio of event membership prior to sampling. (default: 200)}

\item{ndraws}{number of random samples. Runtime is linear in the number of random samples. (default: 200)}
}
\value{
distmat a matrix of estimated FR statistics across multiple samples. Rows and columns are indexed in the order
					in which each sample appears in the input list. For example, a two sample comparison entails rows and
					columns of \emph{distmat} indexed as 1.x follwed by 2.y where x is the Sample 1 population IDs, and
					2.y is the Sample 2 population IDs.
}
\description{
Perform \code{\link{getFRest}} for multiple sample comparisons. For a two-sample comparison of n1 and n2
cell populations, we generate a similarity matrix of dimension (n1+n2)-by-(n1+n2).
}
\author{
Chiaowen Joyce Hsiao \email{joyce.hsiao1@gmail.com}
}

