% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{RunUMAPlayout}
\alias{RunUMAPlayout}
\title{Generate UMAP layout for FLOWMAP graph}
\usage{
RunUMAPlayout(
  graph,
  knn.in,
  file.clusters,
  clustering.var,
  file.name = file.name,
  umap_n_neighbors,
  k,
  umap_n_components,
  mode
)
}
\description{
Generate UMAP layout for FLOWMAP graph
}
