% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-helpers.R
\name{ConstructVarAnnotate}
\alias{ConstructVarAnnotate}
\title{ConstructVarAnnotate}
\usage{
ConstructVarAnnotate(FCS.file.name)
}
\arguments{
\item{FCS.file.name}{The full file path to one FCS file intended for analysis}
}
\value{
List mapping channel names (taken from the "name" field of a
flowFrame) to alternative names from the flowFrame "desc" fields
}
\description{
\code{ConstructVarAnnotate} returns the ???
}
\examples{
ConstructVarAnnotate(FCS.file.name = "Desktop/A.fcs")

\dontrun{
ConstructVarAnnotate()
}
}
