% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-from-FCS.R
\name{DownsampleFCS}
\alias{DownsampleFCS}
\title{SPADE: density-dependent downsampling}
\usage{
DownsampleFCS(fcs.file.names, clustering.var, channel.annotate, channel.remove,
  exclude.pctile = 0.01, target.pctile = 0.99, target.number = NULL,
  target.percent = 0.1, transform = TRUE)
}
\description{
This code was adapted from the spade R package,
available here: \url{https://github.com/nolanlab/spade}.
}
\details{
Specifically, the code in this file is adapted from downsample.R.

This code, authored by M. Linderman, P. Qiu, E. Simonds, Z. Bjornson,
and maintained by Michael Linderman <michael.d.linderman@gmail.com>,
was used under the GNU General Public License v. 2.0
(\url{https://github.com/nolanlab/spade/blob/master/LICENSE}) available
here: \url{https://opensource.org/licenses/GPL-2.0}. In accordance with these
license rules, our code is available under GPL-3.0.
}
