% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FLOWMAP}
\alias{FLOWMAP}
\title{FLOWMAP - generate FLOWMAPR analysis results from FCS files in R}
\usage{
FLOWMAP(mode = c("single", "multi", "one"), files, var.remove, var.annotate,
  clustering.var, cluster.numbers = 100, distance.metric = "manhattan",
  minimum = 2, maximum = 5, per = 1, save.folder = getwd(),
  subsamples = 200, name.sort = TRUE, downsample = FALSE, seed.X = 1,
  savePDFs = TRUE, which.palette = "bluered", exclude.pctile = NULL,
  target.pctile = NULL, target.number = NULL, target.percent = NULL, ...)
}
\arguments{
\item{mode}{FLOWMAPR mode to use in analysis based on starting input,
available options include \code{c("single", "multi", "one")}}

\item{files}{File paths for FCS files to be used or a folder containing
the FCS files to be used in analysis}

\item{var.remove}{Vector naming channels to be removed from all downstream analysis}

\item{var.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{clustering.var}{Vector naming channels to be used to calculate distances/differences
between cells for clustering (if requested) and edge-drawing steps}

\item{cluster.numbers}{A single numeric or a vector of numerics}

\item{distance.metric}{Character \code{c("manhattan", "euclidean")}}

\item{minimum}{Numeric}

\item{maximum}{Numeric}

\item{per}{Numeric}

\item{save.folder}{Directory where all results generated should be saved}

\item{subsamples}{A single numeric or a vector of numerics}

\item{name.sort}{Logical}

\item{downsample}{Logical}

\item{seed.X}{Numeric}

\item{savePDFs}{Logical}

\item{which.palette}{Character \code{c("bluered", "jet", "CB")}}

\item{exclude.pctile}{Optional variable, numeric}

\item{target.pctile}{Optional variable, numeric}

\item{target.number}{Optional variable, numeric}

\item{target.percent}{Optional variable, numeric}
}
\value{
???

\url{http://en.wikipedia.org/}
}
\description{
\code{FLOWMAP} returns the ???
}
\details{
This function ???
}
\examples{
FLOWMAP()
}
