% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-alt.R
\name{FLOWMAPfromDF}
\alias{FLOWMAPfromDF}
\title{FLOWMAPfromDF - generate FLOWMAPR analysis results from
dataframes in R}
\usage{
FLOWMAPfromDF(mode = c("single", "multi", "one"), df, project.name,
  clustering.var, distance.metric = "manhattan", minimum = 2, maximum = 5,
  save.folder = getwd(), time.col.label = "Time",
  condition.col.label = NULL, name.sort = TRUE, clustering = FALSE,
  seed.X = 1, savePDFs = TRUE, which.palette = "bluered",
  cluster.numbers = NULL, cluster.mode)
}
\arguments{
\item{mode}{FLOWMAPR mode to use in analysis based on starting input,
available options include \code{c("single", "multi", "one", "one-special")}}

\item{df}{single dataframe, list of dataframes with each member belonging to single-cell
data from a different timepoint, or a list of lists of dataframes belonging to the same timepoint,
but coming from different conditions, to be used in analysis}

\item{project.name}{Character string to label file output of FLOWMAPR analysis}

\item{clustering.var}{Vector naming channels to be used to calculate distances/differences
between cells for clustering (if requested) and edge-drawing steps}

\item{distance.metric}{Character specifying which metric to use to calculate between-node distances, valid options include \code{c("manhattan", "euclidean")}}

\item{minimum}{Numeric value specifying the minimum number of edges that will be allotted
during each density-dependent edge-building step of the FLOW-MAP graph, default value is
set to \code{2}, no less than 2 is recommended}

\item{maximum}{Numeric value specifying the maximum number of edges that will be allotted
during each density-dependent edge-building step of the FLOW-MAP graph, default value is
set to \code{5}, no less than 3 is recommended}

\item{save.folder}{Directory where all results generated should be saved}

\item{time.col.label}{Character specifying the name of the channel with the time labels for each cell}

\item{condition.col.label}{Character specifying the name of the channel with the condition labels for each cell}

\item{name.sort}{Logical specifying whether to sort FCS file path names alphanumerically or use
them in the order supplied by the user}

\item{clustering}{Logical specifying whether to cluster single-cell data}

\item{seed.X}{Numeric value for the seed to set for reproducible FLOWMAPR runs}

\item{savePDFs}{Logical specifying whether to generate PDFs for the resolved graph with
nodes colored by each parameter}

\item{which.palette}{Optional variable, character specifying which color palette to use
in generated PDFs, valid options include \code{c("bluered", "jet", "CB")}, where \code{"CB"}
is a colorblind-friendly option}

\item{cluster.numbers}{Optional variable, single numeric or a vector of numerics specifying
how many clusters to generate from each separate dataframe}

\item{cluster.mode}{Character specifying which clustering algorithm to use, valid options include \code{c("hclust", "kmeans")}}
}
\value{
the force-directed layout resolved igraph graph object
}
\description{
\code{FLOWMAP} generates FLOWMAPR analysis results from dataframe(s). For more information,
as well as a guide for how to choose the best settings for your analysis, go to
our GitHub repo \url{https://github.com/zunderlab/FLOWMAP/}.
}
