% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-alt-helpers.R
\name{RestructureDF}
\alias{RestructureDF}
\title{Restructure a dataframe into a list of dataframes}
\usage{
RestructureDF(df, time.col.label = "Time", condition.col.label = NULL)
}
\arguments{
\item{df}{a single dataframe containing all single-cell data from all treatments
and conditions (if relevant) to be restructured}

\item{time.col.label}{Character specifying the channel name that should
be used to find the timepoint label for each cell, default is set to \code{"Time"}}

\item{condition.col.label}{Character specifying the channel name that should
be used to find the condition label for each cell, default is set to \code{NULL}}
}
\value{
a list of dataframes containing the single-cell data or if \code{condition.col.label}
is provided (not \code{NULL}), a list of lists containing dataframes of single-cell data
}
\description{
\code{RestructureDF} generates the list of dataframes containing single-cell data
where each consecutive member is a dataframe from a different timepoint, from a single
dataframe where a column designates the timepoints and if relevant, the conditions, of
each individual cell. This function is a pre-processing step before starting a FLOWMAPR
analysis using FLOWMAPfromDF as it only accepts single-cell data in a certain format.
}
\examples{
\dontrun{df <- read.csv(file = "/single-cell-data.csv", header = TRUE, sep = ",")}
\dontrun{RestructureDF(df, time.col.label = "Timepoint", condition.col.label = "Treatment")}
}
