% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{RunUMAPlayout}
\alias{RunUMAPlayout}
\title{Generate UMAP layout for FLOWMAP graph}
\usage{
RunUMAPlayout(graph, file.clusters, file.name = file.name, umap.settings)
}
\description{
Generate UMAP layout for FLOWMAP graph
}
