% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-helpers.R
\name{SuggestClusteringVar}
\alias{SuggestClusteringVar}
\title{SuggestClusteringVar}
\usage{
SuggestClusteringVar(fcs.file.names, mode, var.annotate, var.remove, top.num)
}
\arguments{
\item{fcs.file.names}{A vector of full file paths to the FCS files to be used}

\item{mode}{FLOWMAPR mode to use in analysis based on starting input,
available options include \code{c("single", "multi", "one")}}

\item{var.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{var.remove}{Vector naming channels to be removed from all loaded FCS data}

\item{top.num}{Numeric specifying the number of variables in the vector to be returned}
}
\value{
Vector naming channels that vary the most within and between FCS files
}
\description{
\code{SuggestClusteringVar} returns a vector of user-specified length that contains
suggested names of channels to be used as \code{clustering.var} in FLOWMAPR analysis.
Chosen channels are determined based on analysis of the provided FCS files, as the
variance is calculated within and between timepoints. The top varying markers are noted
and tabulated, and the markers that consistently vary the most are selected.
}
\examples{
fcs.file.names <- c("Desktop/1.fcs", "Desktop/2.fcs")
var.annotate <- ConstructVarAnnotate[1]
var.remove <- c("Channel3", "Channel4")

SuggestClusteringVar(fcs.file.names, mode = "single", var.annotate,
var.remove, top.num = 20)
}
