% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-helpers.R
\name{SuggestVarRemove}
\alias{SuggestVarRemove}
\title{SuggestVarRemove}
\usage{
SuggestVarRemove(var.annotate, var.to.remove = NULL)
}
\arguments{
\item{var.annotate}{List mapping channel names to user-specified names to properly
annotate all FCS file data}

\item{var.to.remove}{Vector of substrings to, default is set to \code{NULL}}
}
\value{
Vector naming channels suggested to be removed from downstream FLOWMAPR analysis
}
\description{
\code{SuggestVarRemove} returns the ???
}
\examples{
var.annotate <- ConstructVarAnnotate("Desktop/A.fcs")
SuggestVarRemove(var.annotate, var.to.remove = "Blank_Channel")
\dontrun{
SuggestVarRemove(var.to.remove)
}
}
